Extract process execution into a class and inject console object to improve testability
