+++ packages/devcmd/src/index.ts
@@ -1,2 +1,2 @@
-export { exec, execParallel } from "./process";
 export { devcmd } from "./devcmd";
+export { exec, execParallel, ProcessExecutor, ProcessInfo } from "./process";
