+++ packages/devcmd/src/process/index.ts
@@ -1,0 +1,4 @@
+import { ProcessExecutor } from "./ProcessExecutor";
+
+export const { exec, execParallel } = new ProcessExecutor(console);
+export { ProcessExecutor, ProcessInfo } from "./ProcessExecutor";
